% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addSentiment.R
\name{addSentiment}
\alias{addSentiment}
\title{Add Sentiment Scores as barchart to current chob}
\usage{
addSentiment(fieldname, on = NA)
}
\arguments{
\item{fieldname}{Specifies fieldname in XTS object which contains sentiment scores}

\item{on}{Screen on which to plot}
}
\description{
Add Sentiment Scores as barchart to current chob
}
\author{
Mario Annau
}
\references{
adapted from Jeffrey A. Ryan's \code{\link{quantmod}} package
}
\seealso{
\code{\link{addSentimentVo}} \code{\link{chartSentiment}}
}

