% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addSentiment.R
\name{addSubjectivity}
\alias{addNeg_Refs_Per_Ref}
\alias{addPolarity}
\alias{addPos_Refs_Per_Ref}
\alias{addSentVol}
\alias{addSenti_Diffs_Per_Ref}
\alias{addSubjectivity}
\title{Add Sentiment Field to chobs
Adds one of the following sentiment fields to the current chob:
Subjectivity, Polarity, Pos_Refs_Per_Ref, Pos_Refs_Per_Ref, Neg_Refs_Per_Ref,
Senti_Diffs_Per_Ref, SentVol}
\usage{
addSubjectivity(..., on = NA, legend = "auto")
}
\description{
Add Sentiment Field to chobs
Adds one of the following sentiment fields to the current chob:
Subjectivity, Polarity, Pos_Refs_Per_Ref, Pos_Refs_Per_Ref, Neg_Refs_Per_Ref,
Senti_Diffs_Per_Ref, SentVol
}
\author{
Mario Annau
}

