% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chartSent}
\alias{chartSent}
\title{Chart Sentiment is a Helper function for \code{\link{addSentiment}}}
\usage{
chartSent(x)
}
\arguments{
\item{x}{chob}
}
\description{
Chart Sentiment is a Helper function for \code{\link{addSentiment}}
}
\author{
Mario Annau
}
\seealso{
\code{\link{addSentiment}}
}

