% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chartSentiment}
\alias{chartSentiment}
\title{Chart Sentiment
Chart (stock) sentiment using quantmods plotting capabilities.
Some hopefully useful presets have been made.}
\usage{
chartSentiment(xts, prefix = "SEMNT", sentname = paste(prefix, "polarity",
  sep = "."), volname = paste(prefix, "vol", sep = "."), volcolor = "blue",
  na.fill = TRUE, omit.na.leading = TRUE, type = "line",
  theme = "white", name, TA = "", postTA = "", ...)
}
\arguments{
\item{xts}{XTS Object Containing at minimum 'Close' and 'Score' fields}

\item{prefix}{prefix of sentiment field, defaults to "SEMNT"}

\item{sentname}{name of sentiment score field, defaults to paste(prefix,"polarity", sep = ".")}

\item{volname}{name of sentiment volume field, defaults to paste(prefix,"volume", sep = ".")}

\item{volcolor}{Color of Sentiment Score Volume Bars, defaults to "blue"}

\item{na.fill}{should parts of timeseries be plotted where \code{sentname} or \code{volname}
in \code{xts} is \code{NA}? Could be useful if Sentiment score/volume have been aggregated or
data is very sparse. Please note that timeseries then becomes compressed, defaults to TRUE}

\item{omit.na.leading}{Omits leading NA's in timeseries, defaults to TRUE}

\item{type}{linetype of Stock Chart, defaults to "line",}

\item{theme}{Chart Theme, defaults to "white"}

\item{name}{Main Title of Sentiment Chart, default is taken from prefix in time series}

\item{TA}{add Technical analysis Indicators to chart at the beginning as string, defaults to ""}

\item{postTA}{add Technical analysis Indicators at the end to the chart, defaults to ""}

\item{...}{additional parameters to chartSeries function call}
}
\description{
Chart Sentiment
Chart (stock) sentiment using quantmods plotting capabilities.
Some hopefully useful presets have been made.
}
\author{
Mario Annau
}
\seealso{
Most presets and optional arguments refer to quantmods excellent
\code{\link{chartSeries}} function
}

