% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getField.VCorpus}
\alias{getField.VCorpus}
\title{Get Meta Data Field in VCorpus as Vector
Get Meta Data Field in VCorpus as Vector}
\usage{
\method{getField}{VCorpus}(x, fieldname)
}
\arguments{
\item{x}{VCorpus object}

\item{fieldname}{character which specifies meta data field name in corpus}
}
\description{
Get Meta Data Field in VCorpus as Vector
Get Meta Data Field in VCorpus as Vector
}
\author{
Mario Annau
}

