% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/score.R
\name{neg_refs_per_ref}
\alias{neg_refs_per_ref}
\alias{neg_refs_per_ref.DocumentTermMatrix}
\alias{neg_refs_per_ref.TermDocumentMatrix}
\title{Calculate neg_refs_per_ref from Matrix}
\usage{
neg_refs_per_ref(x, negative)

\method{neg_refs_per_ref}{TermDocumentMatrix}(x, negative = negterms_GI())

\method{neg_refs_per_ref}{DocumentTermMatrix}(x, ...)
}
\arguments{
\item{x}{TermDocumentMatrix}

\item{negative}{character vector specifying negative terms to be used, defaults to  \code{\link{negterms_GI}}}
}
\description{
Calculate neg_refs_per_ref from Matrix
}
\author{
Mario Annau
}

