% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pos_refs_per_ref}
\alias{pos_refs_per_ref}
\title{Calculate pos_refs_per_ref from Matrix}
\usage{
pos_refs_per_ref(x, positive)
}
\arguments{
\item{x}{TermDocumentMatrix}

\item{positive}{character vector specifying positive terms to be used, defaults to \code{\link{posterms_GI}}}
}
\description{
Calculate pos_refs_per_ref from Matrix
}
\author{
Mario Annau
}

