% Generated by roxygen2 (4.0.1): do not edit by hand
\name{senti_diffs_per_ref}
\alias{senti_diffs_per_ref}
\title{Calculate senti_diffs_per_ref from Matrix}
\usage{
senti_diffs_per_ref(x, positive, negative)
}
\arguments{
\item{x}{TermDocumentMatrix}

\item{positive}{character vector specifying positive terms to be used, defaults to \code{\link{posterms_GI}}}

\item{negative}{character vector specifying negative terms to be used, defaults to\code{\link{negterms_GI}}}
}
\description{
Calculate senti_diffs_per_ref from Matrix
}
\author{
Mario Annau
}

