% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setField.VCorpus}
\alias{setField.VCorpus}
\title{Set Meta Data Field in VCorpus from Vector
Set Meta Data Field in VCorpus from Vector}
\usage{
\method{setField}{VCorpus}(x, fieldname, value)
}
\arguments{
\item{x}{VCorpus object}

\item{fieldname}{character which specifies meta data field name in corpus}

\item{value}{vector which values should be used}
}
\description{
Set Meta Data Field in VCorpus from Vector
Set Meta Data Field in VCorpus from Vector
}
\note{
vector must be of same length as Corpus
}
\author{
Mario Annau
}

