\name{Subj}
\alias{Subj}
\title{Simple XTS transformations for (OHLC) Stock Sentiment XTS Objects...}
\usage{Subj(x)
}
\description{Simple XTS transformations for (OHLC) Stock Sentiment XTS Objects}
\alias{Subj}
\alias{has.Subj}
\alias{Polarity}
\alias{has.Polarity}
\alias{Pos_Refs_Per_Ref}
\alias{has.Pos_Refs_Per_Ref}
\alias{Neg_Refs_Per_Ref}
\alias{has.Neg_Refs_Per_Ref}
\alias{Senti_Diffs_Per_Ref}
\alias{has.Senti_Diffs_Per_Ref}
\alias{SentVol}
\alias{has.SentVol}
\seealso{\code{quantmod}}
\arguments{\item{x}{xts object}
}

