% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/score.R
\name{subjectivity}
\alias{subjectivity}
\alias{subjectivity.DocumentTermMatrix}
\alias{subjectivity.TermDocumentMatrix}
\title{Calculate Polarity from Matrix}
\usage{
subjectivity(x, positive, negative)

\method{subjectivity}{TermDocumentMatrix}(x, positive = posterms_GI(),
  negative = negterms_GI())

\method{subjectivity}{DocumentTermMatrix}(x, ...)
}
\arguments{
\item{x}{TermDocumentMatrix}

\item{positive}{character vector specifying positive terms to be used, defaults to \code{\link{posterms_GI}}}

\item{negative}{character vector specifying negative terms to be used, defaults to\code{\link{negterms_GI}}}
}
\description{
Calculate Polarity from Matrix
}
\author{
Mario Annau
}

