% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OHCLSent.transformations.R
\name{Subj}
\alias{Neg_Refs_Per_Ref}
\alias{Polarity}
\alias{Pos_Refs_Per_Ref}
\alias{SentVol}
\alias{Senti_Diffs_Per_Ref}
\alias{Subj}
\alias{has.Neg_Refs_Per_Ref}
\alias{has.Polarity}
\alias{has.Pos_Refs_Per_Ref}
\alias{has.SentVol}
\alias{has.Senti_Diffs_Per_Ref}
\alias{has.Subj}
\title{Simple XTS transformations for (OHLC) Stock Sentiment XTS Objects}
\usage{
Subj(x)

has.Subj(x, which = FALSE)

Polarity(x)

has.Polarity(x, which = FALSE)

Pos_Refs_Per_Ref(x)

has.Pos_Refs_Per_Ref(x, which = FALSE)

Neg_Refs_Per_Ref(x)

has.Neg_Refs_Per_Ref(x, which = FALSE)

Senti_Diffs_Per_Ref(x)

has.Senti_Diffs_Per_Ref(x, which = FALSE)

SentVol(x, prefix = "SEMNT")

has.SentVol(x, which = FALSE, prefix = "SEMNT")
}
\arguments{
\item{x}{xts object}
}
\description{
Simple XTS transformations for (OHLC) Stock Sentiment XTS Objects
}
\seealso{
\code{quantmod}
}

