\name{addSentiment}
\alias{addSentiment}
\title{Add Sentiment Scores as barchart to current chob...}
\usage{addSentiment(fieldname, on=NA)
}
\description{Add Sentiment Scores as barchart to current chob}
\author{Mario Annau}
\seealso{\code{\link{addSentimentVo}} \code{\link{chartSentiment}}}
\references{adapted from Jeffrey A. Ryan's \code{\link{quantmod}} package}
\arguments{\item{fieldname}{Specifies fieldname in XTS object which contains sentiment scores}
\item{on}{Screen on which to plot}
}

