% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addSentimentVo}
\alias{addSentimentVo}
\title{Add Sentiment Volume as barchart to current chob}
\usage{
addSentimentVo(fieldname, color = "blue")
}
\arguments{
\item{fieldname}{Specifies fieldname in XTS object which contains sentiment volume, default "Volume"}

\item{color}{Color of barchart, default "blue"}
}
\description{
Add Sentiment Volume as barchart to current chob
}
\author{
Mario Annau
}
\seealso{
\code{\link{addSentimentVo}} \code{\link{chartSentiment}}
}

