\name{addSubjectivity}
\alias{addSubjectivity}
\title{Add Sentiment Field to chob...}
\description{Add Sentiment Field to chob
Adds one of the following sentiment fields to the current chob:
Subjectivity, Polarity, Pos_Refs_Per_Ref, Pos_Refs_Per_Ref, Neg_Refs_Per_Ref, 
Senti_Diffs_Per_Ref, SentVol}
\alias{addSubjectivity}
\alias{addPolarity}
\alias{addPos_Refs_Per_Ref}
\alias{addNeg_Refs_Per_Ref}
\alias{addSenti_Diffs_Per_Ref}
\alias{addSentVol}
\author{Mario Annau}

