\docType{data}
\name{dic_gi}
\alias{dic_gi}
\title{This is data to be included in \code{tm.plugin.sentiment}}
\description{
  This is data to be included in \code{tm.plugin.sentiment}
}
\author{
  Mario Annau \email{mario.annau@gmail.com}
}
\references{
  \url{http://www.wjh.harvard.edu/~inquirer/}
}
\keyword{data}
\keyword{datasets}
\keyword{general}
\keyword{inquirer}

