\name{getField.VCorpus}
\alias{getField.VCorpus}
\title{Get Meta Data Field in VCorpus as Vector...}
\usage{getField.VCorpus(x, fieldname)
}
\description{Get Meta Data Field in VCorpus as Vector
Get Meta Data Field in VCorpus as Vector}
\author{Mario Annau}
\arguments{\item{x}{VCorpus object}
\item{fieldname}{character which specifies meta data field name in corpus}
}

