% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neg_refs_per_ref}
\alias{neg_refs_per_ref}
\title{Calculate neg_refs_per_ref from Matrix}
\usage{
neg_refs_per_ref(x, negative)
}
\arguments{
\item{x}{TermDocumentMatrix}

\item{negative}{character vector specifying negative terms to be used, defaults to  \code{\link{negterms_GI}}}
}
\description{
Calculate neg_refs_per_ref from Matrix
}
\author{
Mario Annau
}

