% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/score.R
\name{senti_diffs_per_ref}
\alias{senti_diffs_per_ref}
\alias{senti_diffs_per_ref.DocumentTermMatrix}
\alias{senti_diffs_per_ref.TermDocumentMatrix}
\title{Calculate senti_diffs_per_ref from Matrix}
\usage{
senti_diffs_per_ref(x, positive, negative)

\method{senti_diffs_per_ref}{TermDocumentMatrix}(x, positive = posterms_GI(),
  negative = negterms_GI())

\method{senti_diffs_per_ref}{DocumentTermMatrix}(x, ...)
}
\arguments{
\item{x}{TermDocumentMatrix}

\item{positive}{character vector specifying positive terms to be used, defaults to \code{\link{posterms_GI}}}

\item{negative}{character vector specifying negative terms to be used, defaults to\code{\link{negterms_GI}}}
}
\description{
Calculate senti_diffs_per_ref from Matrix
}
\author{
Mario Annau
}

