\name{setField.VCorpus}
\alias{setField.VCorpus}
\title{Set Meta Data Field in VCorpus from Vector...}
\usage{setField.VCorpus(x, fieldname, value)
}
\description{Set Meta Data Field in VCorpus from Vector
Set Meta Data Field in VCorpus from Vector}
\author{Mario Annau}
\note{vector must be of same length as Corpus}
\arguments{\item{x}{VCorpus object}
\item{fieldname}{character which specifies meta data field name in corpus}
\item{value}{vector which values should be used}
}

