\name{Subj}
\alias{has.Neg_Refs_Per_Ref}
\alias{has.Polarity}
\alias{has.Pos_Refs_Per_Ref}
\alias{has.Senti_Diffs_Per_Ref}
\alias{has.SentVol}
\alias{has.Subj}
\alias{Neg_Refs_Per_Ref}
\alias{Polarity}
\alias{Pos_Refs_Per_Ref}
\alias{Senti_Diffs_Per_Ref}
\alias{SentVol}
\alias{Subj}
\title{Simple XTS transformations for (OHLC) Stock Sentiment XTS Objects}
\usage{
  Subj(x)
}
\arguments{
  \item{x}{xts object}
}
\description{
  Simple XTS transformations for (OHLC) Stock Sentiment XTS
  Objects
}
\seealso{
  \code{quantmod}
}

