% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subjectivity}
\alias{subjectivity}
\title{Calculate Polarity from Matrix}
\usage{
subjectivity(x, positive, negative)
}
\arguments{
\item{x}{TermDocumentMatrix}

\item{positive}{character vector specifying positive terms to be used, defaults to \code{\link{posterms_GI}}}

\item{negative}{character vector specifying negative terms to be used, defaults to\code{\link{negterms_GI}}}
}
\description{
Calculate Polarity from Matrix
}
\author{
Mario Annau
}

